Fix hardcoded display widths and improve timing overflow handling

+++ src/main.cpp
@@ -129,7 +129,8 @@ void loop() {
     }
     
     // Return to scanning display after detection timeout
-    if (millis() - lastDisplayUpdate > DISPLAY_UPDATE_INTERVAL) {
+    // Cast handles millis() overflow correctly (wraps every ~49 days)
+    if ((unsigned long)(millis() - lastDisplayUpdate) > DISPLAY_UPDATE_INTERVAL) {
         displayScanning(SCAN_FREQUENCY);
         lastDisplayUpdate = millis();
     }
